+++ BayPassTests/APIs/HereTests.swift
@@ -64,23 +64,22 @@ class HereTests: XCTestCase {
 
     func testGetAgencyFromStationId() {
         let ex = expectation(description: "Here for getting Agency")
-        let stationId: Int = 718_310_131
-        let time = "2019-06-24T08%3A00%3A00"
+        let stationId: Int = 718_610_044
         var result: Agency?
 
-        here.getAgencyFromStationId(stationId: stationId, time: time) { resp in
+        here.getAgencyFromStationId(stationId: stationId) { resp in
             result = resp
             ex.fulfill()
         }
         wait(for: [ex], timeout: 5)
-        XCTAssertEqual(result!, Agency.BART)
+        XCTAssertEqual(result!, Agency.CalTrain)
     }
 
     func testGetAgencyFromStationIdFail() {
         let ex = expectation(description: "Here for getting Agency")
         var result: Agency?
 
-        here.getAgencyFromStationId(stationId: 10, time: "") { resp in
+        here.getAgencyFromStationId(stationId: 10) { resp in
             result = resp
             ex.fulfill()
         }
@@ -92,11 +91,10 @@ class HereTests: XCTestCase {
         let ex = expectation(description: "Here for getting Station Ids")
         let center = CLLocationCoordinate2D(latitude: 37.5032238, longitude: -121.9434281)
         let radius = 1500
-        let time = "2019-03-24T08%3A00%3A00"
 
         var results: [Station]? = [Station]()
 
-        here.getStationsNearby(center: center, radius: radius, max: 1, time: time) { resp in
+        here.getStationsNearby(center: center, radius: radius, max: 1) { resp in
             results = resp
             ex.fulfill()
         }
