+++ BayPassTests/Controller/ClipperManagerTests.swift
@@ -31,7 +31,7 @@ class ClipperManagerTests: XCTestCase {
     }
 
     func testGetValidPasses() {
-        let calTrain = Agency(name: "CalTrain", routes: [], icon: #imageLiteral(resourceName: "CalTrain"))
+        let calTrain = Agency(rawValue: "CT")!
         let testCard = ClipperCard(number: 9_999_999_999, cashValue: 0.0, passes: [Pass(name: "Monthly", duration: DateInterval(start: Date(), duration: 36000), price: 45.0, validOnAgency: calTrain)])
         ClipperManager.shared.setClipperCard(card: testCard)
         let validPasses = ClipperManager.shared.getValidPasses()
