+++ BayPassTests/Controller/ClipperViewControllerTests.swift
@@ -15,7 +15,7 @@ class ClipperViewControllerTests: XCTestCase {
     override func setUp() {
         let calTrain = Agency(rawValue: "CT")
         let testCard = ClipperCard(number: 9_999_999_999, cashValue: 0.0, passes: [Pass(name: "Monthly", duration: DateInterval(start: Date(), duration: 36000), price: 45.0, validOnAgency: calTrain ?? Agency.zero)])
-        clipperManager.setClipperCard(card: testCard)
+        ClipperManager.shared.setClipperCard(card: testCard)
 
         UIApplication.shared.keyWindow!.rootViewController = vc
         XCTAssertNotNil(UIApplication.shared.keyWindow?.rootViewController)
